refactor(network): remove old upload from everywhere
