+++ src/services/network/index.ts
@@ -1,3 +1,1 @@
-export { uploadFile } from './upload';
 export { downloadFile } from './download';
-export { deleteFile } from './delete';
